+++ pkg/front_end/test/extensions/data/ambiguous/lib1.dart
@@ -7,6 +7,8 @@
   AmbiguousExtension2,
   UnambiguousExtension1,
   async.dart.FutureExtensions,
+  collection.dart.IterableExtensions,
+  collection.dart.NullableIterableExtensions,
   core.dart.DateTimeCopyWith,
   core.dart.EnumByName,
   core.dart.EnumName]*/