+++ pkg/front_end/test/extensions/data/as_show/lib.dart
@@ -5,6 +5,8 @@
 /*library: scope=[
   Extension1,
   async.dart.FutureExtensions,
+  collection.dart.IterableExtensions,
+  collection.dart.NullableIterableExtensions,
   core.dart.DateTimeCopyWith,
   core.dart.EnumByName,
   core.dart.EnumName]*/