+++ pkg/front_end/test/extensions/data/as_show/origin.dart
@@ -5,6 +5,8 @@
 /*library: scope=[
   Extension2,
   async.dart.FutureExtensions,
+  collection.dart.IterableExtensions,
+  collection.dart.NullableIterableExtensions,
   core.dart.DateTimeCopyWith,
   core.dart.EnumByName,
   core.dart.EnumName]*/