+++ pkg/front_end/test/extensions/data/extension_on_type_variable.dart
@@ -5,6 +5,8 @@
 /*library: scope=[
   GeneralGeneric,
   async.dart.FutureExtensions,
+  collection.dart.IterableExtensions,
+  collection.dart.NullableIterableExtensions,
   core.dart.DateTimeCopyWith,
   core.dart.EnumByName,
   core.dart.EnumName]*/