+++ pkg/front_end/test/extensions/data/named_declarations.dart
@@ -8,6 +8,8 @@
   B3,
   B4,
   async.dart.FutureExtensions,
+  collection.dart.IterableExtensions,
+  collection.dart.NullableIterableExtensions,
   core.dart.DateTimeCopyWith,
   core.dart.EnumByName,
   core.dart.EnumName]*/