+++ pkg/front_end/test/extensions/data/patching/origin.dart
@@ -6,6 +6,8 @@
   Extension,
   GenericExtension,
   async.dart.FutureExtensions,
+  collection.dart.IterableExtensions,
+  collection.dart.NullableIterableExtensions,
   core.dart.DateTimeCopyWith,
   core.dart.EnumByName,
   core.dart.EnumName]*/