+++ pkg/front_end/test/extensions/data/prefix/lib3.dart
@@ -5,6 +5,8 @@
 /*library: scope=[
   ShownExtension3,
   async.dart.FutureExtensions,
+  collection.dart.IterableExtensions,
+  collection.dart.NullableIterableExtensions,
   core.dart.DateTimeCopyWith,
   core.dart.EnumByName,
   core.dart.EnumName]*/