+++ pkg/front_end/test/extensions/data/reexport/lib2.dart
@@ -6,6 +6,8 @@
   ClashingExtension,
   UniqueExtension2,
   async.dart.FutureExtensions,
+  collection.dart.IterableExtensions,
+  collection.dart.NullableIterableExtensions,
   core.dart.DateTimeCopyWith,
   core.dart.EnumByName,
   core.dart.EnumName]*/