+++ pkg/front_end/test/extensions/data/reexport/main.dart
@@ -4,6 +4,8 @@
 
 /*library: scope=[
   async.dart.FutureExtensions,
+  collection.dart.IterableExtensions,
+  collection.dart.NullableIterableExtensions,
   core.dart.DateTimeCopyWith,
   core.dart.EnumByName,
   core.dart.EnumName,