+++ pkg/front_end/test/extensions/data/type_variables.dart
@@ -7,6 +7,8 @@
   A3,
   A4,
   async.dart.FutureExtensions,
+  collection.dart.IterableExtensions,
+  collection.dart.NullableIterableExtensions,
   core.dart.DateTimeCopyWith,
   core.dart.EnumByName,
   core.dart.EnumName]*/